Add language model
In this model, we can get the available language in the app. That is,
english and dutch.

A migration script is created to create the languages table and add a
foreign key in irregular_verbs table to link with languages table.

A seed is also created to create the both languages available.