Add category feature for vocabularies
This model will be use to classify the vocabularies by category.
A category will be linked to a language and one or more vocabularies.
For instance, in english, we could decide to define add phrasal verbs as
vocabularies.

A migration script is written to create categories table.

A seed is also created to define the default categories for both
languages.