Review validators
Before this commit, the validators are only useful for the create record
because when we update a record we only have the attribtute to update
for record. So, all attributes are optionals when we want to update a
record.

This commit renames all existing validators to Create{Name}Validator
since those validators are only useful for the create record.
Moreover, it create validators for the update record.