Add show method in controller and change orderby
This commit adds the show method in Category, Vocabularies and
IrregularVerbs controllers to search a record in their respective model.
It also changes the orderBy when we returns the vocabularies or
irregular verbs in controllers of for those records. Before this commit,
we shuffle the list of records before give it to the frontend
application. Now, it returns the list ordered by the 'word' column if
we are in the Vocabularies controller and by 'infinitive' column if we
are in the IrregularVerbs controller.