Fix the irregular verb validators
Before this commit, when the frontend call the api to create or update
an irregular verb, the validator in the api rejected the verb given in
the request. Because the verb is given in snake_case and not in
camelCase.

This commit fixes the validators to take the verb in snake_case format
to correctly create/update an irregular verb.