Convert the irregular verb given in parameter for store method
Before this commit, when the user wants to add an irregular verb he has
an error 400 (Bad Request), because the api does not have a field called
`past_simple2` in the IrregularVerb model. In fact, it is `past_simple_2`.

This commit converts the verb given in parameter after validating it to
correctly store it in database.